mysql/json:refactor
