+++ src/util/codec/mysql/json/comparison.rs
@@ -11,8 +11,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// FIXME(shirly): remove following later
-#![allow(dead_code)]
 use std::cmp::Ordering;
 use std::f64;
